Made Several Functionality Updates.
Completed changes:
1. Change colour of buttons (e.g. Add To Queue green when on the block,
Grey when not).
2. Hide The Draft Number At The Top Of The Budgets Pane By Making It
Same Colour As Background.
3. Add Budget to Budgets Pane And Update Max Bid.
4. Update "Add To Queue" text to "Add To Block".
5. Add logic to allow coach to specify how much they put a player on
the block for initially (to stop issues when budgets are low).
6. Update Emails So That They Are Team Names.
7. Add “Next” button only to the admin coaches page.
8. Headers don’t line up with data in Data Tables.
9. Add ability to sort by Number Taken to Drafted Players list.

Note: Potentially introduced a strange bug that stops the user from
logging out using the Log Out button however this seems to be
intermittent. I don’t believe that this was in the previous version.