Removed Need For Admin To Be Logged In To Progress Draft.
Removed the need for the admin coach to be logged in in order for the draft to be automated. This was achieved by adding a clientList variable that tracks the websockets connections and disconnections. We then compare the individual users socketID with the first socketID that is in the clientList array and run the draft() function off of their client to progress the draft.