Minor Layout Updates Following Last Nights' Successful Test (ROLLBACK POINT IF REQUIRED).
1. Moved injured players like Docherty down the list.
2. Added an underline for the current on the block coach that stays there during bidding and is cleared when the draft is complete.
3. Changed the select drop down on the Draft page back to the standard setup so that they are easier to view.
4. Changed the budget field on the Create A Draft page to a drop down so that it that only takes numbers. Budgets are now restricted to $100 increments between $100 and $1000.

THIS IS OUR ROLLBACK POINT IF WE NEED TO ROLLBACK AS THIS HAS BEEN TESTED WITH 8 COACHES IN A DRAFT AND WITH TOMMY IN ANOTHER TIMEZONE AND BOTH WORKED. Only slight issue was a small discrepancy in the countdown timers for Sam twice (and potentially Ben once) that corrected itself. I am still looking into this issue.